reader: optimization
As soon as we have read all the fields we can
immediately create a user/opening.

This was previously done when we encountered a "begging" of user/opening
where we checked if our fields (secttor/username) were not empty.
This also led to having to handle the "special"
last case, where no user/opening follows.

With this commit all the code to create the user is in one place and
no special cases are required.